requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.12
langchain-experimental>=0.0.43
chromadb>=0.4.22
openai>=1.6.1
tiktoken>=0.5.2
aiohttp>=3.9.1
requests>=2.31.0
langgraph>=0.0.20
python-magic-bin>=0.4.14
tree-sitter>=0.20.4
radon>=6.0.1
pylint>=3.0.3
mistune>=3.0.2
networkx>=3.2.1
docstring-parser>=0.15
sentence-transformers>=2.2.2
nltk>=3.8.1
